checking INT MAX and INT min

+++ libft.c
@@ -23,11 +23,11 @@ static int	ft_isspace(const char c)
 		return (0);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	sign = 1;
@@ -45,8 +45,6 @@ int	ft_atoi(const char *str)
 		result = result * 10 + (str[i] - '0');
 		i++;
 	}
-	// if (result > 2147483647 || result < -2147483648)
-	// 	return (0);
 	return (result * sign);
 }
